+++ ServersideQoL.AutoProcess/CHANGELOG.md
@@ -1,4 +1,4 @@
-### v2.0.15
+### v2.1.0
 - Added option `FeedFromContainersRangeSignPrefix` [#270](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/270) 
 - Fixed an issue that caused range configuration with ContainerSigns to not work [#273](https://github.com/ArgusMagnus/ValheimServersideQoL/pull/273) (contributed by [Lempac](https://github.com/Lempac))
 - Added option `FeedFromContainersExcludeOre` to exclude certain ore items (e.g. `FineWood` in kilns) from being fed to smelters from containers