+++ ServersideQoL.TameAssist/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v2.1.0
+- Added options to feed tames from containers (`FeedFromContainers`) [#15](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/15)
+
 ### v2.0.11
 - Support for the new core options `UnifiedConfig` and `ConfigPerWorld`